add `Consumer::receive_all`
Support for batched reads is not uniform among the various backends we
have.

The in-memory and rabbitmq impls differ from the rest as these consumers
are channel/stream based, with no explicit option for consuming messages
in batches.

Using `tokio::time::timeout` for the first item, then looping to
opportunistically try to fill up the batch with any remaining
_already ready_ messages, we can emulate a similar behavior as the other
backends.

In all cases, receive_all should:
- not block longer (+/-) than the specified deadline, even if no items
  are available.
- not linger waiting for _more items_ once it as received _any_.

Additionally, a new bit of config has been exposed for rabbitmq so callers
can specify a prefetch count. This didn't seem to have any impact on the
tests, but seems good to have.