Upgrade redis-rs/bb8 again
With the reconfigured clustered redis healthcheck,
I believe we're safe to upgrade redis again.

Note: This version of redis-rs imposes a default 1-second
timeout on establishing connections to cluster nodes. I don't
see any reason at the moment to change this, but we may need
to consider exposing this as a configurable parameter in the future.