Add redis eviction-policy check/warning
We want to warn users if queues are at potential risk of deletion
in low-memory situations. Add a startup check that will log a
warning if the eviction policy can't be determined or if an unsafe
eviction policy is found.

Note: The `warn` is likely to fire all the time on cloud providers,
which tend to disable the `CONFIG` command entirely.