Add redis sentinel support
There are a few breaking changes made to support this.
Most notably, the an optional `SentinelConfig` struct
has been added to the RedisConfig object. It seems like
obfuscating the fields here might be a good idea, but
that's not how we've been rolling thus far with the
config objects.

Also, the `RedisConnection` trait has been redone
to take a `RedisConfig` instead of a DSN in order
to build the connection. Sentinel needs it owing
to the large number of potential config params needed
for Sentinel, and I think this is probably a better
approach anyway.

The base redis tests (well, most of them) have also
been updated to support testing against either
base Redis or Sentinel. The `rstest` crate has been
added in order to be able to parameterize the tests
sanely. We also use this crate for tests in `redis-rs`,
so I think it's a no-brainer to add here.